Add README and extend functionality in main.py
This commit introduces a README file for the Muslim API project. The README provides an overview of the project and instructions on how to run it. Additionally, the main.py file has been enhanced with new API endpoints. One endpoint returns all data from the 'asmaul-husna.json' file, while another can return specific data based on id. If the id does not exist, an HTTPException with status code 404 is raised. These changes are made to increase user comprehension and server functionality.