"Delete doa.py file
The doa.py file has been removed. The file was used to store data and metadata for each type of Doa. However, this data has been moved to a database which significantly improves the way we handle, manage, and retrieve this information. Therefore, the doa.py file was no longer required and has been deleted to declutter the codebase."